requirements.txt
streamlit
openai>=1.3.0
python-dotenv
